Fix https://issues.apache.org/bugzilla/show_bug.cgi?id=57581 Change statistics byte counter in coyote Request object to be long to allow values above 2Gb. (RequestInfo.getRequestBytesReceived() already exposes it as long to JMX.)
I am changing the setter method as well. It is never called, removed from Tomcat 7 onwards.

git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc6.0.x/trunk@1660792 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/coyote/Request.java
@@ -139,7 +139,7 @@ public final class Request {
     private Response response;
     private ActionHook hook;
 
-    private int bytesRead=0;
+    private long bytesRead=0;
     // Time of the request - usefull to avoid repeated calls to System.currentTime
     private long startTime = 0L;
     private int available = 0;
@@ -521,11 +521,11 @@ public final class Request {
         return reqProcessorMX;
     }
 
-    public int getBytesRead() {
+    public long getBytesRead() {
         return bytesRead;
     }
 
-    public void setBytesRead(int bytesRead) {
+    public void setBytesRead(long bytesRead) {
         this.bytesRead = bytesRead;
     }
 }
